Switch mutex implementation to match errno's implementation
