+++ libc/cmake/modules/LLVMLibCCompileOptionRules.cmake
@@ -110,6 +110,10 @@ function(_get_compile_options_from_config output_var)
     list(APPEND config_options "-DLIBC_ERRNO_MODE=${LIBC_CONF_ERRNO_MODE}")
   endif()
 
+  if(LIBC_CONF_THREAD_MODE)
+    list(APPEND config_options "-DLIBC_THREAD_MODE=${LIBC_CONF_THREAD_MODE}")
+  endif()
+
   set(${output_var} ${config_options} PARENT_SCOPE)
 endfunction(_get_compile_options_from_config)
