--- libc/src/__support/threads/baremetal/mutex.h
@@ -1,32 +1,0 @@
-//===--- Implementation of a mutex class for baremetal ----------*- C++ -*-===//
-//
-// Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
-// See https://llvm.org/LICENSE.txt for license information.
-// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
-//
-//===----------------------------------------------------------------------===//
-
-#ifndef LLVM_LIBC_SRC___SUPPORT_THREADS_BAREMETAL_MUTEX_H
-#define LLVM_LIBC_SRC___SUPPORT_THREADS_BAREMETAL_MUTEX_H
-
-#include "src/__support/macros/attributes.h"
-#include "src/__support/macros/config.h"
-#include "src/__support/threads/mutex_common.h"
-
-namespace LIBC_NAMESPACE_DECL {
-
-/// Implementation of a simple passthrough mutex which guards nothing. For
-/// single threaded processors, this is the implementation.
-struct Mutex {
-  LIBC_INLINE constexpr Mutex(bool, bool, bool, bool) {}
-
-  LIBC_INLINE MutexError lock() { return MutexError::NONE; }
-  LIBC_INLINE MutexError unlock() { return MutexError::NONE; }
-  LIBC_INLINE MutexError reset() { return MutexError::NONE; }
-};
-
-// TODO: add multithreading support here
-
-} // namespace LIBC_NAMESPACE_DECL
-
-#endif // LLVM_LIBC_SRC___SUPPORT_THREADS_BAREMETAL_MUTEX_H
