--- libc/src/__support/threads/gpu/CMakeLists.txt
@@ -1,5 +1,0 @@
-add_header_library(
-  mutex
-  HDRS
-    mutex.h
-)
